requirements.txt
asn1crypto==1.4.0
astroid==2.5.6
attrs==20.3.0
aws-cdk.assets==1.101.0
aws-cdk.aws-apigateway==1.101.0
aws-cdk.aws-applicationautoscaling==1.101.0
aws-cdk.aws-autoscaling==1.101.0
aws-cdk.aws-autoscaling-common==1.101.0
aws-cdk.aws-autoscaling-hooktargets==1.101.0
aws-cdk.aws-batch==1.101.0
aws-cdk.aws-certificatemanager==1.101.0
aws-cdk.aws-cloudformation==1.101.0
aws-cdk.aws-cloudfront==1.101.0
aws-cdk.aws-cloudwatch==1.101.0
aws-cdk.aws-codebuild==1.101.0
aws-cdk.aws-codecommit==1.101.0
aws-cdk.aws-codedeploy==1.101.0
aws-cdk.aws-codeguruprofiler==1.101.0
aws-cdk.aws-codepipeline==1.101.0
aws-cdk.aws-codepipeline-actions==1.101.0
aws-cdk.aws-cognito==1.101.0
aws-cdk.aws-dynamodb==1.101.0
aws-cdk.aws-ec2==1.101.0
aws-cdk.aws-ecr==1.101.0
aws-cdk.aws-ecr-assets==1.101.0
aws-cdk.aws-ecs==1.101.0
aws-cdk.aws-efs==1.101.0
aws-cdk.aws-eks==1.101.0
aws-cdk.aws-elasticloadbalancing==1.101.0
aws-cdk.aws-elasticloadbalancingv2==1.101.0
aws-cdk.aws-events==1.101.0
aws-cdk.aws-events-targets==1.101.0
aws-cdk.aws-globalaccelerator==1.101.0
aws-cdk.aws-iam==1.101.0
aws-cdk.aws-kinesis==1.101.0
aws-cdk.aws-kinesisfirehose==1.101.0
aws-cdk.aws-kms==1.101.0
aws-cdk.aws-lambda==1.101.0
aws-cdk.aws-logs==1.101.0
aws-cdk.aws-route53==1.101.0
aws-cdk.aws-route53-targets==1.101.0
aws-cdk.aws-s3==1.101.0
aws-cdk.aws-s3-assets==1.101.0
aws-cdk.aws-sam==1.101.0
aws-cdk.aws-secretsmanager==1.101.0
aws-cdk.aws-servicediscovery==1.101.0
aws-cdk.aws-signer==1.101.0
aws-cdk.aws-sns==1.101.0
aws-cdk.aws-sns-subscriptions==1.101.0
aws-cdk.aws-sqs==1.101.0
aws-cdk.aws-ssm==1.101.0
aws-cdk.aws-stepfunctions==1.101.0
aws-cdk.aws-stepfunctions-tasks==1.101.0
aws-cdk.cloud-assembly-schema==1.101.0
aws-cdk.core==1.101.0
aws-cdk.custom-resources==1.101.0
aws-cdk.cx-api==1.101.0
aws-cdk.lambda-layer-awscli==1.101.0
aws-cdk.lambda-layer-kubectl==1.101.0
aws-cdk.pipelines==1.101.0
aws-cdk.region-info==1.101.0
awswrangler==2.7.0
beautifulsoup4==4.9.3
boto3==1.17.61
botocore==1.20.61
cattrs==1.5.0
certifi==2020.12.5
chardet==4.0.0
constructs==3.3.75
et-xmlfile==1.1.0
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
isort==5.8.0
jmespath==0.10.0
jsii==1.29.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pg8000==1.18.0
pluggy==0.13.1
publication==0.0.3
py==1.10.0
pyarrow==3.0.0
pylint==2.8.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
redshift-connector==2.0.878
requests==2.25.1
s3transfer==0.4.2
scramp==1.2.2
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.12.1
zipp==3.4.1